+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-08 16:02:36
-// Commit: feat(core): add bi-directional event bus handler for core
+// Updated at: 2026-04-09 11:15:11
+// Commit: fix(core): fix memory leak by cancelling core subscription
 
 class coreComponentState {
-  final String commitRef = '193';
-  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
-  final DateTime timestamp = DateTime.parse('2026-04-08T16:02:36+05:00');
+  final String commitRef = '213';
+  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
+  final DateTime timestamp = DateTime.parse('2026-04-09T11:15:11+05:00');
 }